+++ jib-core/src/test/java/com/google/cloud/tools/jib/frontend/JavaDockerContextGeneratorTest.java
@@ -151,8 +151,9 @@ public class JavaDockerContextGeneratorTest {
     String dockerfile =
         new JavaDockerContextGenerator(mockJavaLayerConfigurations)
             .setBaseImage(expectedBaseImage)
-            .setJvmFlags(expectedJvmFlags)
-            .setMainClass(expectedMainClass)
+            .setEntrypoint(
+                JavaEntrypointConstructor.makeDefaultEntrypoint(
+                    expectedJvmFlags, expectedMainClass))
             .setJavaArguments(expectedJavaArguments)
             .setExposedPorts(exposedPorts)
             .setLabels(expectedLabels)